requirements-google.txt
# Optional: Google Calendar sync for generate_todo_pages.py --google-calendar
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
